# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MOTHERSHIP Multi-Agent Governance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rules:
    get:
      operationId: listRules
      tags: [queen]
      summary: List all governance rules
      responses:
        "200":
          description: Rules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovernanceRule"
    post:
      operationId: createRule
      tags: [queen]
      summary: Create a new governance rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleBody"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceRule"
...
components:
  schemas:
...
    GovernanceRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - severity
        - active
        - createdAt
    CreateRuleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
      required:
        - name
        - description
        - severity
...